tech: retirer codes inutiles
