--- src/components/custom/Coach/CoachContact.vue
@@ -1,21 +1,0 @@
-<template>
-  <div class="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
-    <div class="fr-col-md-8">
-      <h2 class="fr-h4 text--white">J'agis est un nouveau service qui continue de s’améliorer</h2>
-      <p class="text--white fr-text--xl fr-mb-0">
-        Posez vos questions, partagez vos retours et suggérez de nouvelles fonctionnalités que vous aimeriez
-        voir sur <span class="text--italic">J'agis</span>.
-      </p>
-    </div>
-    <div class="fr-col-md-4">
-      <a
-        href="https://thread-origami-ae6.notion.site/130576eceac1804b8863d874f7f1fd1e?pvs=105"
-        class="fr-btn fr-btn--icon-left fr-icon-survey-line border--white"
-        target="_blank"
-      >
-        Donner mon avis
-      </a>
-    </div>
-  </div>
-</template>
-<script setup lang="ts"></script>
