--- src/domaines/serviceRecherche/catalogue/adapters/serviceRechercheViewModel.ts
@@ -1,19 +1,0 @@
-interface AsideServiceViewModel {
-  nom: string;
-  description: string;
-  url: string;
-  urlLabel: string;
-  logo: string;
-  screenshot: string;
-}
-
-interface CategoriesViewModel {
-  code: string;
-  label: string;
-  estLaCategorieParDefaut: boolean;
-}
-
-export interface ServiceRechercheViewModelBase {
-  aside: AsideServiceViewModel;
-  categories: CategoriesViewModel[];
-}
