--- src/domaines/serviceRecherche/catalogue/ports/serviceRecherche.presenter.ts
@@ -1,5 +1,0 @@
-import { ServicesRecherche } from '@/domaines/serviceRecherche/catalogue/recupererServicesRechercheParThematique.usecase';
-
-export interface ServiceRecherchePresenter {
-  presente(serviceRecherche: ServicesRecherche): void;
-}
