--- src/domaines/serviceRecherche/catalogue/serviceRecherche.ts
@@ -1,9 +1,0 @@
-interface ServiceRechercheCategorie {
-  code: string;
-  label: string;
-  estLaCategorieParDefaut: boolean;
-}
-
-export interface ServiceRechercheBase {
-  categories: ServiceRechercheCategorie[];
-}
