+++ src/domaines/serviceRecherche/suggestionServiceViewModel.ts
@@ -16,3 +16,33 @@ export interface SuggestionServiceViewModel {
   };
   categories?: string;
 }
+
+interface AsideServiceViewModel {
+  nom: string;
+  description: string;
+  url: string;
+  urlLabel: string;
+  logo: string;
+  screenshot: string;
+}
+
+interface CategoriesViewModel {
+  code: string;
+  label: string;
+  estLaCategorieParDefaut: boolean;
+}
+
+export interface ServiceRechercheViewModelBase {
+  aside: AsideServiceViewModel;
+  categories: CategoriesViewModel[];
+}
+
+interface ServiceRechercheCategorie {
+  code: string;
+  label: string;
+  estLaCategorieParDefaut: boolean;
+}
+
+export interface ServiceRechercheBase {
+  categories: ServiceRechercheCategorie[];
+}
