+++ tests/serviceRecherche/longueVieAuxObjets/recupererServiceLongueVieAuxObjets.usecase.spec.ts
@@ -91,7 +91,7 @@ describe("Fichier de tests concernant la récuperation de service d'un recherche
                 },
                 tag: {
                   label: 'À 300 m',
-                  style: 'background--caramel text--background-caramel',
+                  style: 'fr-tag--custom-bleu',
                 },
                 to: {
                   name: RouteServiceName.LONGUE_VIE_AUX_OBJETS_DETAIL,
@@ -125,7 +125,7 @@ describe("Fichier de tests concernant la récuperation de service d'un recherche
                 nombreMiseEnFavoris: 0,
                 tag: {
                   label: 'À 2,1 km',
-                  style: 'background--caramel text--background-caramel',
+                  style: 'fr-tag--custom-bleu',
                 },
                 titre: 'titre 3',
                 to: {
